backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.MD"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
python-jose = "^3.3.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
httpx = "^0.27.0"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
pytest = "^8.0.2"
bandit = "^1.7.7"
mypy = "^1.9.0"
pylint = "^3.1.0"
anyio = "^4.3.0"
types-python-jose = "^3.3.4.20240106"
pylint-pydantic = "^0.3.2"
freezegun = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Mypy rules
# Based on this https://careers.wolt.com/en/blog/tech/professional-grade-mypy-configuration
[tool.mypy]
plugins = [
  "pydantic.mypy"
]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = true

[tool.black]
line-length = 100

[tool.pylint]
load-plugins = "pylint_pydantic"
min-similarity-lines=10

[tool.poetry.scripts]
start-dev = "app.main:start_dev"
format = "scripts:format_code"
format-check = "scripts:format_code_check"
lint = "scripts:lint_code"
type-check = "scripts:type_check"
sec-check = "scripts:security_check"
test = "scripts:run_tests"